chore: split Panel cypress tests into multiple files
